Hackily avoid another kind of rounding error
The main fix here is rounding intersection points to integers, if
they're already /very very close/ to integers.  This is super shady, but
it fixes some nasty algorithm-breaking behavior in the common case of
Doom maps, where all the coordinates are integers.

This code can now successfully process MAP01 and dilated shells for
every single line in it!