Fix calculation of holes
The original code relied heavily on always tracing counter-clockwise, so
that contours were guaranteed to be traced from the outside inwards.  I
broke that contract when I extended the tracing to handle each face of
an edge separately, because now the inner CCW contour is traced before
an outer CW contour around the same edges.

But I only care about holes versus not-holes; anything else should act
like a top-level contour, not an island inside another hole.  With that
in mind, I can delete most of this logic and only handle the hole case,
which DOES have the guarantee that the outer contour has been traced
first.